Move tests for getting/setting Judgment properties into their own file
